test(platform): split quiescence helpers from runner
